pyproject.toml
[tool.poetry]
name = "demo-app"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.taskipy.tasks]
pre_test = "task format && task check"
test = "pytest"

format = "task format_autoflake && task format_pyupgrade && task format_black && task format_isort"
format_autoflake = "autoflake --in-place --remove-all-unused-imports $(find src tests -name '*.py')"
format_black = "black -l 79 src tests"
format_isort = "isort --profile black -l 79 src tests"
format_pyupgrade = "pyupgrade $(find src tests -name '*.py')"

check = "task check_mypy"
check_mypy = "mypy src tests"

[tool.mypy]
check_untyped_defs = true
disallow_any_generics = true
disallow_untyped_defs = true
ignore_missing_imports = true
plugins = "strawberry.ext.mypy_plugin"
python_version = 3.11
show_column_numbers = true
show_error_codes = true
strict_optional = true
warn_return_any = true
warn_unreachable = true
warn_unused_configs = true

[tool.pytest.ini_options]
addopts = "--strict-markers --strict-config -ra -vv -s --cov=src --cov-report=html"
testpaths = "tests"
pythonpath = "src"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.103.1"
uvicorn = {extras = ["standard"], version = "^0.23.2"}
strawberry-graphql = {extras = ["fastapi"], version = "^0.205.0"}
supabase = "^1.0.4"
sqlalchemy = "^2.0.20"
asyncpg = "^0.28.0"
strawberry-sqlalchemy-mapper = "^0.2.0"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}

[tool.poetry.group.dev.dependencies]
taskipy = "^1.12.0"
pytest = "^7.4.1"
pytest-mock = "^3.11.1"
pytest-cov = "^4.1.0"
mypy = "^1.5.1"
black = "^23.7.0"
flake8 = "^6.1.0"
isort = "^5.12.0"
autoflake = "^2.2.1"
pyupgrade = "^3.10.1"
types-requests = "^2.31.0.2"
httpx = "^0.24.1"
pytest-asyncio = "^0.21.1"
alembic = "^1.12.0"
pytest-postgresql = "^5.0.0"
psycopg = {extras = ["binary"], version = "^3.1.10"}

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
